requirements.txt
appdirs==1.4.3
asgiref==3.2.3
backcall==0.1.0
cycler==0.10.0
decorator==4.4.1
Django==3.0.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
kiwisolver==1.1.0
matplotlib==3.1.2
nose==1.3.7
numpy==1.18.0
pandas==0.25.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
ply==3.11
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
Pyomo==5.6.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyUtilib==5.7.3
scipy==1.4.1
six==1.13.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
python-dotenv
psycopg2
